+++ libs/common/src/lib/interfaces/portfolio-details.interface.ts
@@ -16,6 +16,15 @@ export interface PortfolioDetails {
   filteredValueInBaseCurrency?: number;
   filteredValueInPercentage: number;
   holdings: { [symbol: string]: PortfolioPosition };
+  platforms: {
+    [id: string]: {
+      balance: number;
+      currency: string;
+      name: string;
+      valueInBaseCurrency: number;
+      valueInPercentage?: number;
+    };
+  };
   summary: PortfolioSummary;
   totalValueInBaseCurrency?: number;
 }